+++ src/services/ldap/index.js
@@ -222,12 +222,13 @@ module.exports = function LDAPService() {
 		 * @return {Promise[Array[Object]]} resolves with array of objects
 		 * matching the query, rejects with error otherwise
 		 */
-		searchCollection(config, searchString, options = {}, rawAttributes = []) {
+		searchCollection(config, searchString, options = {}, rawAttributes = [], firstPageOnly = false) {
 			// Paging to avoid 'max size limit exceeded' issue
 			const optionsWithPaging = {
 				...options,
 				paged: {
 					pageSize: 100,
+					pagePause: firstPageOnly,
 				},
 			};
 
@@ -240,6 +241,11 @@ module.exports = function LDAPService() {
 								reject(err);
 							}
 							res.on('error', reject);
+							res.on('page', (result) => {
+								if (firstPageOnly) {
+									resolve(objects);
+								}
+							});
 							res.on('searchEntry', (entry) => {
 								const result = entry.object;
 								rawAttributes.forEach((element) => {
@@ -295,6 +301,17 @@ module.exports = function LDAPService() {
 		 */
 		getUsers(config, school) {
 			return getLDAPStrategy(app, config).getUsers(school);
+		}
+
+		/**
+		 * Returns all users at a school on the LDAP server
+		 * @param {LdapConfig} config the ldapConfig
+		 * @param {School} school the school object
+		 * @return {Promise[Object]} resolves with all user objects or rejects
+		 * with error
+		 */
+		verifyConfig(config, school) {
+			return getLDAPStrategy(app, config).verifyConfig(school);
 		}
 
 		/**